+++ smarts/core/chassis.py
@@ -40,7 +40,12 @@ from smarts.core.utils.bullet import (
     JointInfo,
     JointState,
 )
-from smarts.core.utils.math import radians_to_vec, yaw_from_quaternion
+from smarts.core.utils.math import (
+    min_angles_difference_signed,
+    radians_to_vec,
+    vec_to_radians,
+    yaw_from_quaternion,
+)
 from smarts.core.utils.pybullet import bullet_client as bc
 
 with pkg_resources.path(models, "vehicle.urdf") as path:
@@ -111,7 +116,7 @@ class Chassis:
         raise NotImplementedError
 
     @property
-    def yaw_rate(self):
+    def yaw_rate(self) -> float:
         raise NotImplementedError
 
     def inherit_physical_values(self, other: "Chassis"):
@@ -151,10 +156,14 @@ class BoxChassis(Chassis):
         bullet_client.setCollisionFilterGroupMask(
             self._bullet_body._bullet_id, -1, 0x0, 0x0
         )
+        self._pose = None
         self.control(pose, speed)
         self._client = bullet_client
 
-    def control(self, pose: Pose, speed: float):
+    def control(self, pose: Pose, speed: float, dt: float = 0):
+        if self._pose:
+            self._last_heading = self._pose.heading
+        self._last_dt = dt
         self._pose = pose
         self._speed = speed
         self._bullet_constraint.move_to(pose)
@@ -177,11 +186,22 @@ class BoxChassis(Chassis):
 
     @property
     def speed(self):
+        # in m/s
         return self._speed
 
     @property
     def velocity_vectors(self):
-        return (None, None)
+        # linear_velocity in m/s, angular_velocity in rad/s
+        vh = radians_to_vec(self.pose.heading)
+        if self._speed:
+            linear_velocity = np.array((vh[0] * self._speed, vh[1] * self._speed, 0.0))
+        else:
+            linear_velocity = None
+        angular_velocity = np.array((0.0, 0.0, 0.0))
+        if self._last_dt > 0:
+            angular_velocity = vh - radians_to_vec(self._last_heading)
+            angular_velocity = np.append(angular_velocity / self._last_dt, 0.0)
+        return (linear_velocity, angular_velocity)
 
     @speed.setter
     def speed(self, speed: float = None):
@@ -196,7 +216,11 @@ class BoxChassis(Chassis):
         return None
 
     @property
-    def yaw_rate(self):
+    def yaw_rate(self) -> float:
+        # in rad/s
+        if self._last_dt > 0:
+            delta = min_angles_difference_signed(self._pose.heading, self._last_heading)
+            return delta / self._last_dt
         return None
 
     def inherit_physical_values(self, other: Chassis):
@@ -408,7 +432,7 @@ class AckermannChassis(Chassis):
 
     @property
     def velocity_vectors(self):
-        """Linear velocity vector is in km/h and Angular veclocity is in Rad/sec"""
+        """Linear velocity vector is in m/s and Angular veclocity is in Rad/sec"""
         linear_velocity, angular_velocity = np.array(
             self._client.getBaseVelocity(self._bullet_id)
         )
@@ -424,10 +448,10 @@ class AckermannChassis(Chassis):
             self.control(throttle=1)
 
     @property
-    def yaw_rate(self):
-        """Returns rotational speed in rad/sec."""
+    def yaw_rate(self) -> float:
+        """Returns 2-D rotational speed in rad/sec."""
         _, velocity_rotational = np.array(self._client.getBaseVelocity(self._bullet_id))
-        return velocity_rotational
+        return vec_to_radians(velocity_rotational[:2])
 
     @property
     def longitudinal_lateral_speed(self):