+++ smarts/core/controllers/imitation_controller.py
@@ -19,10 +19,15 @@
 # THE SOFTWARE.
 import math
 import numpy as np
-from dataclasses import replace
 
 from smarts.core.chassis import AckermannChassis, BoxChassis
-from smarts.core.utils.math import fast_quaternion_from_angle, radians_to_vec
+from smarts.core.coordinates import Pose
+from smarts.core.utils.math import (
+    fast_quaternion_from_angle,
+    min_angles_difference_signed,
+    radians_to_vec,
+    vec_to_radians,
+)
 
 
 class ImitationController:
@@ -32,43 +37,51 @@ class ImitationController:
         if isinstance(action, (int, float)):
             # special case:  setting the initial speed
             if isinstance(chassis, BoxChassis):
-                vehicle.control(vehicle.pose, action)
+                vehicle.control(vehicle.pose, action, dt)
             elif isinstance(chassis, AckermannChassis):
                 chassis.speed = action  # hack that calls control internally
             return
 
         assert isinstance(action, (list, tuple)) and len(action) == 2
-        # acceleration in m/s^2, angluar_velocity in rad/s
+        # acceleration is vector in m/s^2, angluar_velocity is scalar in rad/s
         acceleration, angular_velocity = action
+        target_heading = (vehicle.heading + angular_velocity * dt) % (2 * math.pi)
 
         if isinstance(chassis, BoxChassis):
-            target_speed = vehicle.speed + acceleration * dt
-            target_heading = (vehicle.heading + angular_velocity * dt) % (2 * math.pi)
-            speed_x, speed_y = radians_to_vec(target_heading)
-            new_position = [
-                vehicle.position[0] + dt * speed_x * target_speed,
-                vehicle.position[1] + dt * speed_y * target_speed,
-                vehicle.position[2],
-            ]
-            new_pose = replace(
-                vehicle.pose,
+            speed_x, speed_y = radians_to_vec(target_heading) * vehicle.speed
+            speed_x += acceleration[0] * dt
+            speed_y += acceleration[1] * dt
+            new_position = np.array(
+                (
+                    vehicle.position[0] + dt * speed_x,
+                    vehicle.position[1] + dt * speed_y,
+                    vehicle.position[2],
+                )
+            )
+            new_pose = Pose(
                 orientation=fast_quaternion_from_angle(target_heading),
                 position=new_position,
             )
-            vehicle.control(new_pose, target_speed)
+            vehicle.control(new_pose, np.linalg.norm((speed_x, speed_y)), dt)
 
         elif isinstance(chassis, AckermannChassis):
+            acc = np.linalg.norm(acceleration)
+            acc_angle = min_angles_difference_signed(
+                target_heading, vec_to_radians(acceleration[:2])
+            )
             mass = chassis.mass_and_inertia[0]  # in kg
-            if acceleration >= 0:
+            wheel_radius = chassis.wheel_radius
+            if abs(acc_angle) <= math.pi / 2:
                 # necessary torque is N*m = kg*m*acceleration
-                torque_ratio = mass / (4 * chassis.wheel_radius * chassis.max_torque)
-                throttle = np.clip(acceleration * torque_ratio, 0, 1)
+                torque_ratio = mass / (4 * wheel_radius * chassis.max_torque)
+                throttle = np.clip(acc * torque_ratio, 0, 1)
                 brake = 0
             else:
                 throttle = 0
                 # necessary torque is N*m = kg*m*acceleration
-                torque_ratio = mass / (4 * chassis.wheel_radius * chassis.max_btorque)
-                brake = np.clip(-acceleration * torque_ratio, 0, 1)
+                torque_ratio = mass / (4 * wheel_radius * chassis.max_btorque)
+                brake = np.clip(acc * torque_ratio, 0, 1)
+
             steering = np.clip(dt * -angular_velocity * chassis.steering_ratio, -1, 1)
             vehicle.control(throttle=throttle, brake=brake, steering=steering)